+++ R/CoxModel.R
@@ -109,7 +109,7 @@ fitCoxModel<- function(population, plpData, param, search='adaptive',
   }
   
   #get prediction on test set:
-  prediction <- predict.plp(plpModel=list(model = modelTrained),
+  prediction <- predict_plp(plpModel=list(model = modelTrained),
                             population = population, 
                             plpData = plpData)