+++ tests/testthat/helper-objects.R
@@ -1,7 +1,29 @@
+
+# Download the PostreSQL driver ---------------------------
+# If DATABASECONNECTOR_JAR_FOLDER exists, assume driver has been downloaded
+jarFolder <- Sys.getenv("DATABASECONNECTOR_JAR_FOLDER", unset = "")
+if (jarFolder == "") {
+  tempJarFolder <- tempfile("jdbcDrivers")
+  dir.create(tempJarFolder)
+  Sys.setenv("DATABASECONNECTOR_JAR_FOLDER" = tempJarFolder)
+  downloadJdbcDrivers("postgresql")
+  
+  withr::defer({
+    unlink(tempJarFolder, recursive = TRUE, force = TRUE)
+    Sys.unsetenv("DATABASECONNECTOR_JAR_FOLDER")
+  }, testthat::teardown_env())
+}
+
+
 # this files contains the objects used in the tests:
 travis <- T
-saveLoc <- tempfile()
 
+saveLoc <- tempfile("saveLoc")
+dir.create(saveLoc)
+
+# configure and activate python
+PatientLevelPrediction::configurePython(envname = 'r-reticulate', envtype = "conda")
+PatientLevelPrediction::setPythonEnvironment(envname = 'r-reticulate', envtype = "conda")
 
 #@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@
 # simulated data Tests
@@ -25,6 +47,9 @@ plpData3$covariateData$covariatesOld <- plpData3$covariateData$covariates
 plpData3$covariateData$covariates <- plpData3$covariateData$covariatesOld %>% 
   dplyr::filter(.data$covariateId <= 20 ) %>%
   dplyr::mutate(timeId = 1+.data$covariateId*.data$rowId%%3)
+
+sampleSize4 <- 10000
+plpData4 <- simulatePlpData(plpDataSimulationProfile, n = sampleSize4)
 
 
 # POPULATION
@@ -54,12 +79,27 @@ population2 <- createStudyPopulation(plpData2,
                                      riskWindowEnd = 365,
                                      endAnchor = 'cohort start')
 
+population4 <- createStudyPopulation(plpData4,
+                                     outcomeId = 2,
+                                     firstExposureOnly = FALSE,
+                                     washoutPeriod = 0,
+                                     removeSubjectsWithPriorOutcome = FALSE,
+                                     priorOutcomeLookback = 99999,
+                                     requireTimeAtRisk = T,
+                                     minTimeAtRisk=10,
+                                     riskWindowStart = 0,
+                                     startAnchor = 'cohort start',
+                                     riskWindowEnd = 365,
+                                     endAnchor = 'cohort start')
+
 
 # MODEL SETTINGS
 lrSet <- setLassoLogisticRegression()
 gbmSet <- setGradientBoostingMachine(ntrees = 50, maxDepth = 3, learnRate = 0.01, seed = 1)
 knnSet <- setKNN(k=100, indexFolder = file.path(saveLoc,"knn"))
 rfSet2 <- setRandomForest(mtries = -1,ntrees = 10, maxDepth = 2, varImp = F, seed=1)
+surv <- PatientLevelPrediction::setCoxModel(variance = 0.01, seed = 1)
+
 
 
 # RUNPLP - LASSO LR
@@ -72,6 +112,18 @@ plpResult <- runPlp(population = population,
                     savePlpPlots = F, 
                     analysisId = 'lrTest',
                     saveDirectory =  saveLoc)
+
+# RUNPLP - survival Cox
+plpResult2 <- runPlp(population = population,
+                    plpData = plpData, 
+                    modelSettings = surv, 
+                    savePlpData = F, 
+                    savePlpResult = F, 
+                    saveEvaluation = F, 
+                    savePlpPlots = F, 
+                    analysisId = 'survTest',
+                    saveDirectory =  saveLoc)
+
 
 
 # learningCurve 
@@ -150,5 +202,4 @@ plpResultReal <- runPlp(population = populationReal,
                         savePlpPlots = F, 
                         analysisId = 'gbmReal',
                         saveDirectory =  saveLoc)
-